refactor(ErrorResponse, GlobalExceptionHandler): merge medication and user exceptions
