+++ src/main/java/pildonitas/pdnt/exceptions/ErrorResponse.java
@@ -1,27 +1,42 @@
 package pildonitas.pdnt.exceptions;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
 
 import java.time.LocalDateTime;
 
-@Getter
-@JsonPropertyOrder({"timestamp", "status", "error", "message", "path"})
 public class ErrorResponse {
+    private String message;
+    private int status;
+    private String errorCode;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    private final LocalDateTime timestamp;
-    private final int status;
-    private final String error;
-    private final String message;
-    private final String path;
+    private LocalDateTime timestamp;
+    private String path;
 
-    public ErrorResponse(HttpStatus status, String message, String path) {
-        this.timestamp = LocalDateTime.now();
-        this.status = status.value();
-        this.error = status.name();
+    public ErrorResponse(String message, int status, String errorCode, String path) {
         this.message = message;
+        this.status = status;
+        this.errorCode = errorCode;
+        this.timestamp = LocalDateTime.now();
         this.path = path;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+
+    public String getErrorCode() {
+        return errorCode;
+    }
+
+    public LocalDateTime getTimestamp() {
+        return timestamp;
+    }
+
+    public String getPath() {
+        return path;
     }
 }
