+++ src/main/java/pildonitas/pdnt/exceptions/GlobalExceptionHandler.java
@@ -3,43 +3,70 @@ package pildonitas.pdnt.exceptions;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.HashMap;
-import java.util.Map;
+import pildonitas.pdnt.medication.exceptions.custom_exceptions.MedicationNotFoundException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
     @ExceptionHandler(EntityNotFoundException.class)
     public ResponseEntity<ErrorResponse> handleEntityNotFoundException(EntityNotFoundException exception, HttpServletRequest req) {
-        HttpStatus status = HttpStatus.NOT_FOUND;
-        String controllerPath = String.valueOf(req.getRequestURL());
-        ErrorResponse errorResponse = new ErrorResponse(status, exception.getMessage(), controllerPath);
-
-        return new ResponseEntity<>(errorResponse, status);
+        ErrorResponse error = new ErrorResponse(
+                exception.getMessage(),
+                HttpStatus.NOT_FOUND.value(),
+                "ENTITY_NOT_FOUND",
+                req.getRequestURI()
+        );
+        return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
     }
 
     @ExceptionHandler(EntityAlreadyExistsException.class)
     public ResponseEntity<ErrorResponse> handleEntityAlreadyExistsException(EntityAlreadyExistsException exception, HttpServletRequest req) {
-        HttpStatus status = HttpStatus.CONFLICT;
-        String controllerPath = String.valueOf(req.getRequestURL());
-        ErrorResponse errorResponse = new ErrorResponse(status, exception.getMessage(), controllerPath);
+        ErrorResponse error = new ErrorResponse(
+                exception.getMessage(),
+                HttpStatus.CONFLICT.value(),
+                "ENTITY_ALREADY_EXISTS",
+                req.getRequestURI()
+        );
+        return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
+    }
 
-        return new ResponseEntity<>(errorResponse, status);
+    @ExceptionHandler(MedicationNotFoundException.class)
+    public ResponseEntity<ErrorResponse> MedicationNotFound(MedicationNotFoundException exception, HttpServletRequest request) {
+        ErrorResponse error = new ErrorResponse(
+                exception.getMessage(),
+                HttpStatus.NOT_FOUND.value(),
+                "Medication_NOT_FOUND",
+                request.getRequestURI()
+        );
+        return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<Map<String, String>> handleValidationExceptions(MethodArgumentNotValidException exception) {
-        Map<String, String> errors = new HashMap<>();
+    public ResponseEntity<ErrorResponse> handleValidation(MethodArgumentNotValidException exception, HttpServletRequest request) {
+        String message = exception.getBindingResult().getFieldErrors().stream()
+                .findFirst()
+                .map(error -> error.getField() + ": " + error.getDefaultMessage())
+                .orElse("Invalid input");
 
-        exception.getBindingResult().getAllErrors().forEach((error) -> {
-            String fieldName = ((FieldError) error).getField();
-            String errorMessage = error.getDefaultMessage();
-            errors.put(fieldName, errorMessage);
-        });
-        return new ResponseEntity<>(errors, HttpStatus.BAD_REQUEST);
+        ErrorResponse error = new ErrorResponse(
+                message,
+                HttpStatus.BAD_REQUEST.value(),
+                "VALIDATION_ERROR",
+                request.getRequestURI()
+        );
+        return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ErrorResponse> handleGenericError(Exception ex, HttpServletRequest request) {
+        ErrorResponse error = new ErrorResponse(
+                "Internal server error",
+                HttpStatus.INTERNAL_SERVER_ERROR.value(),
+                "INTERNAL_ERROR",
+                request.getRequestURI()
+        );
+        return new ResponseEntity<>(error, HttpStatus.INTERNAL_SERVER_ERROR);
     }
 }
