--- src/main/java/pildonitas/pdnt/medication/exceptions/ErrorResponse.java
@@ -1,42 +1,0 @@
-package pildonitas.pdnt.medication.exceptions;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-
-import java.time.LocalDateTime;
-
-public class ErrorResponse {
-    private String message;
-    private int status;
-    private String errorCode;
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    private LocalDateTime timestamp;
-    private String path;
-
-    public ErrorResponse(String message, int status, String errorCode, String path) {
-        this.message = message;
-        this.status = status;
-        this.errorCode = errorCode;
-        this.timestamp = LocalDateTime.now();
-        this.path = path;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-
-    public String getErrorCode() {
-        return errorCode;
-    }
-
-    public LocalDateTime getTimestamp() {
-        return timestamp;
-    }
-
-    public String getPath() {
-        return path;
-    }
-}
