--- src/main/java/pildonitas/pdnt/medication/exceptions/GlobalExceptionHandler.java
@@ -1,49 +1,0 @@
-package pildonitas.pdnt.medication.exceptions;
-
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import pildonitas.pdnt.medication.exceptions.custom_exceptions.MedicationNotFoundException;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(MedicationNotFoundException.class)
-    public ResponseEntity<ErrorResponse> MedicationNotFound(MedicationNotFoundException exception, HttpServletRequest request) {
-        ErrorResponse error = new ErrorResponse(
-                exception.getMessage(),
-                HttpStatus.NOT_FOUND.value(),
-                "Medication_NOT_FOUND",
-                request.getRequestURI()
-        );
-        return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ErrorResponse> handleValidation(MethodArgumentNotValidException exception, HttpServletRequest request) {
-        String message = exception.getBindingResult().getFieldErrors().stream()
-                .findFirst()
-                .map(error -> error.getField() + ": " + error.getDefaultMessage())
-                .orElse("Invalid input");
-
-        ErrorResponse error = new ErrorResponse(
-                message,
-                HttpStatus.BAD_REQUEST.value(),
-                "VALIDATION_ERROR",
-                request.getRequestURI()
-        );
-        return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<ErrorResponse> handleGenericError(Exception ex, HttpServletRequest request) {
-        ErrorResponse error = new ErrorResponse(
-                "Internal server error",
-                HttpStatus.INTERNAL_SERVER_ERROR.value(),
-                "INTERNAL_ERROR",
-                request.getRequestURI()
-        );
-        return new ResponseEntity<>(error, HttpStatus.INTERNAL_SERVER_ERROR);
-}
